fix compatibility with old style search forms
Older search forms used 'id' to pass the search query. This is no longer
the case since Greebo which has a dedicated 'q' parameter. However some
templates use their own form (like bootstrap3) and have not adapted to
the new parameter, yet. This results in an empty query and also in an
empty aggregation result.

This patch does a fallback to $ID if no query can be found in 'q' and
also gracefully handles missing aggregation results.